Sort active nodes by last uptime proof.
Since HF19, there is no longer a concept of a per node reward height, so
we need a new sort criterion for active nodes. I believe the most
interesting sort is that by most recent uptime proof.

At the top of the list will appear nodes whose proof is overdue
(i.e. > 1:00:00 ago). This should never be more than a few seconds, so
one can see at a glance any nodes that are ripe for being moved to the
Inactive/Decommissioned Service Nodes section the next time they are
tested by the network. This also makes the list much more dynamic to
watch.

Since the Last Reward Height column is now redundant, we can replace it
with something more interesting, such as the Active Since Height.